+++ atlasdb-tests-shared/src/test/java/com/palantir/atlasdb/transaction/impl/SnapshotTransactionTest.java
@@ -26,12 +26,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
-import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 
 import java.math.BigInteger;
@@ -52,7 +47,6 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicLong;
 import java.util.stream.Collectors;
 
 import org.apache.commons.lang3.mutable.MutableInt;
@@ -914,74 +908,6 @@ public class SnapshotTransactionTest extends AtlasDbTestCase {
     }
 
     @Test
-    public void committedWritesAreAddedToSweepQueue() {
-        List<WriteInfo> table1Writes = ImmutableList.of(
-                WriteInfo.of(Cell.create("a".getBytes(), "b".getBytes()), false, 2L),
-                WriteInfo.of(Cell.create("a".getBytes(), "c".getBytes()), false, 2L),
-                WriteInfo.of(Cell.create("a".getBytes(), "d".getBytes()), true, 2L),
-                WriteInfo.of(Cell.create("b".getBytes(), "d".getBytes()), false, 2L));
-        List<WriteInfo> table2Writes = ImmutableList.of(
-                WriteInfo.of(Cell.create("w".getBytes(), "x".getBytes()), false, 2L),
-                WriteInfo.of(Cell.create("y".getBytes(), "z".getBytes()), false, 2L),
-                WriteInfo.of(Cell.create("z".getBytes(), "z".getBytes()), true, 2L));
-
-        AtomicLong startTs = new AtomicLong(0);
-        txManager.runTaskWithRetry(txn -> {
-            table1Writes.forEach(write -> {
-                put(txn, TABLE1, write);
-            });
-            table2Writes.forEach(write -> {
-                put(txn, TABLE2, write);
-            });
-            return null;
-        });
-
-        verify(sweepQueue).enqueue(eq(TABLE1), eq(table1Writes));
-        verify(sweepQueue).enqueue(eq(TABLE2), eq(table2Writes));
-    }
-
-    @Test
-    public void noWritesAddedToSweepQueueOnConflict() {
-        Cell cell = Cell.create("foo".getBytes(), "bar".getBytes());
-        byte[] value = new byte[1];
-
-        Transaction t1 = txManager.createNewTransaction();
-        Transaction t2 = txManager.createNewTransaction();
-
-        t1.put(TABLE, ImmutableMap.of(cell, value));
-        t2.put(TABLE, ImmutableMap.of(cell, new byte[1]));
-
-        t1.commit();
-        verify(sweepQueue).enqueue(any(), any());
-
-        try {
-            t2.commit();
-            fail();
-        } catch (TransactionConflictException e) {
-            // expected
-        }
-
-        verifyNoMoreInteractions(sweepQueue);
-    }
-
-    @Test
-    public void noWritesAddedToSweepQueueOnException() {
-        Cell cell = Cell.create("foo".getBytes(), "bar".getBytes());
-        byte[] value = new byte[1];
-
-        try {
-            txManager.runTaskWithRetry(txn -> {
-                txn.put(TABLE, ImmutableMap.of(cell, value));
-                throw new RuntimeException("test");
-            });
-        } catch (RuntimeException e) {
-            // expected
-        }
-
-        verifyZeroInteractions(sweepQueue);
-    }
-
-    @Test
     public void getRowsColumnRangesReturnsInOrderInCaseOfAbortedTxns() {
         byte[] row = "foo".getBytes();
         Cell firstCell = Cell.create(row, "a".getBytes());
@@ -1095,7 +1021,7 @@ public class SnapshotTransactionTest extends AtlasDbTestCase {
 
 
     private void put(Transaction txn, TableReference table, WriteInfo write) {
-        if (write.isTombstone()) {
+        if (write.writeRef().isTombstone()) {
             txn.delete(table, ImmutableSet.of(write.cell()));
         } else {
             txn.put(table, ImmutableMap.of(write.cell(), new byte[1]));